list_gemini_models.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()

api_key = os.getenv("GEMINI_API_KEY")
if not api_key:
    print("Error: GEMINI_API_KEY not found in .env")
    exit(1)

genai.configure(api_key=api_key)

print("Listing Gemini Models:")
for m in genai.list_models():
    if 'gemini' in m.name:
        print(f"- {m.name}")

