requirements.txt
Flask
flask-cors
python-dotenv
supabase
psycopg2-binary
pydantic==2.8.2
pydantic-core==2.20.1
numpy
scipy
scikit-learn
torch
torchvision
opencv-python
Pillow
pdfplumber
pypdfium2
easyocr
google-generativeai>=0.7.0